requirements.txt
Django==1.4.3
PyYAML==3.10
cssselect==0.7.1
django-haystack==1.2.7
lxml==3.0.1
nltk==2.0.4
psycopg2==2.4.6
pyquery==1.2.4
wsgiref==0.1.2
